+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- write dependencies on modify.
 - do not write bit.json's `misc` and `lang` properties if the default value is presented.
 - send correct error message when there is invalid inline id (wip).
 - add bind command (which calls the driver bind command).